+++ softMocksConfig/factory.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
  * @author Mougrim <rinat@mougrim.ru>
  */
 
-namespace Mougrim\XdebugProxy;
+namespace Mougrim\XdebugProxy\softMocksConfig;
 
 use Mougrim\XdebugProxy\Factory\SoftMocksFactory;