+++ src/LoggerFormatter.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Mougrim\XdebugProxy;
 
 use Monolog\Formatter\LineFormatter;
+
 use function is_scalar;
-use function /** @noinspection ForgottenDebugOutputInspection */ var_export;
+use function var_export;
 
 /**
  * @author Mougrim <rinat@mougrim.ru>
- * @psalm-suppress PropertyNotSetInConstructor
  */
 class LoggerFormatter extends LineFormatter
 {