+++ tests/bootstrap.php
@@ -9,4 +9,4 @@ namespace Tests\Mougrim\XdebugProxy;
 
 use function dirname;
 
-require dirname(__DIR__).'/vendor/autoload.php';
+require dirname(__DIR__) . '/vendor/autoload.php';